# public/api-spec.yaml
openapi: 3.0.3
info:
  title: AS Operadora API
  version: 1.0.0
  description: Backend unificado Web + Móvil
servers:
  - url: https://app.asoperadora.com
    description: Producción
paths:
  /api/auth/login:
    post:
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
                device_fingerprint:
                  type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
  /api/auth/refresh:
    post:
      summary: Refresh access token
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshResponse'
  /api/auth/logout:
    post:
      summary: Logout (invalidate session)
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                session_token:
                  type: string
                user_id:
                  type: integer
      responses:
        '200':
          description: OK
  /api/bookings:
    get:
      summary: List bookings
      parameters:
        - in: query
          name: userId
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: offset
          schema:
            type: integer
        - in: query
          name: view
          schema:
            type: string
            enum: [compact]
      responses:
        '200':
          description: OK
    post:
      summary: Create booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBookingRequest'
      responses:
        '201':
          description: Created
  /api/hotels:
    get:
      summary: Search hotels
      parameters:
        - in: query
          name: city
          schema:
            type: string
        - in: query
          name: checkin
          schema:
            type: string
        - in: query
          name: checkout
          schema:
            type: string
        - in: query
          name: guests
          schema:
            type: integer
        - in: query
          name: view
          schema:
            type: string
            enum: [compact]
      responses:
        '200':
          description: OK
  /api/search:
    get:
      summary: Unified search
      parameters:
        - in: query
          name: type
          schema:
            type: string
            enum: [hotel, flight]
        - in: query
          name: city
          schema:
            type: string
        - in: query
          name: origin
          schema:
            type: string
        - in: query
          name: destination
          schema:
            type: string
        - in: query
          name: view
          schema:
            type: string
            enum: [compact]
      responses:
        '200':
          description: OK
components:
  schemas:
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          nullable: true
        error:
          nullable: true
        meta:
          nullable: true
        timestamp:
          type: string
    LoginResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                user:
                  type: object
                accessToken:
                  type: string
                refreshToken:
                  type: string
    RefreshResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                accessToken:
                  type: string
    CreateBookingRequest:
      type: object
      properties:
        type:
          type: string
        service_name:
          type: string
        total_price:
          type: number
        currency:
          type: string
        status:
          type: string
        payment_status:
          type: string
        details:
          type: object
        user_id:
          type: integer
        tenant_id:
          type: integer